+++ wtframework/wtf/web/page.py
@@ -28,15 +28,18 @@ import time
 
 
 class PageObject(object):
-    '''
-    Baseclass for PageObjects.
+    """Base class for PageObjects.  The goal of PageObjects is to allow a logical 
+    encapsulation of page recognition, mapping of page elements, and exposing logical 
+    services provided by the page on a higher level for high level tests to consume.
     
     Basic Usage:
-    1) define '_validate_page' method.  This method will check to make sure 
+
+    #. define '_validate_page(webdriver)' method.  This method will check to make sure 
        we are on the correct page.
-    2) define '__cmp__' method to allow page ranking when there are multiple matches 
+    #. define '__cmp__' method to allow page ranking when there are multiple matches 
        to the same page to disambiguate which page should take precedence.
-    '''
+
+    """
     __metaclass__ = abc.ABCMeta #needed to make this an abstract class in Python 2.7
 
     # Webdriver associated with this instance of the PageObject
@@ -45,11 +48,13 @@ class PageObject(object):
 
 
     def __init__(self, webdriver, *args, **kwargs):
-        '''
-        Constructor
-        @param webdriver: Instance of Selenium WebDriver
-        @type webdriver: WebDriver
-        '''
+        """Constructor.  It's better to not call this directly, instead use PageFactory 
+        to instantiate PageObjects.
+        
+        Ars: 
+            webdriver (Webdriver): Selenium Webdriver instance.
+
+        """
         try:
             config_reader=kwargs['config_reader']
         except KeyError:
@@ -80,24 +85,35 @@ class PageObject(object):
 
     @abc.abstractmethod
     def _validate_page(self, webdriver):
-        """
-        Perform checks to validate this page is the correct target page.
+        """Perform checks to validate this page is the correct target page.
         
-        All pages must implement this method.
-        
-        @raise InvalidPageError: Raised when we try to assign the wrong page 
-        to this page object.  This exception should be raised when a page match 
-        fails.  Any other exception type would be consider a code failure.
+        All PageObjects must implement this method.
+
+        Args:
+            webdriver (Webdriver) : instance of Selenium Webdriver
+        Raises:
+            InvalidPageError: Raised when we try to assign the wrong page 
+            to this page object.  This exception should be raised when a page match 
+            fails.  Any other exception type would be consider a code failure.
+            
         """
         return
 
 
     @classmethod
     def create_page(cls, webdriver=None, **kwargs):
-        """
-        Class method short cut to call PageFactory on itself.
-        @param webdriver: WebDriver to associate with this page.
-        @type webdriver: WebDriver
+        """Class method short cut to call PageFactory on itself.  Use it to instantiate 
+        this PageObject using a webdriver.
+        
+        Args:
+            webdriver (Webdriver): Instance of Selenium Webdriver.
+        
+        Returns:
+            PageObject
+        
+        Raises:
+            InvalidPageError
+
         """
         if not webdriver:
             webdriver = WTF_WEBDRIVER_MANAGER.get_driver()
@@ -106,10 +122,18 @@ class PageObject(object):
 
     #Magic methods for enabling comparisons.
     def __cmp__(self, other):
-        """
-        Override this to implement PageObject ranking.  This is used by PageObjectFactory
+        """Override this to implement PageObject ranking.  This is used by PageObjectFactory
         when it finds multiple pages that qualify to map to the current page.  The 
         PageObjectFactory will check which page object is preferable.
+        
+        Args:
+            other (PageObject) : Other page object to compare it against.
+        
+        Returns:
+            int ::
+                postive number - this page is prioritized over the other page.
+                negative number - this page is lower in priority than the other page.
+
         """
         if not isinstance(other, PageObject):
             # By default page object will rank itself over non page objects.
@@ -125,31 +149,49 @@ class InvalidPageError(Exception):
 
 
 class PageFactory():
-    "Page Factory class for constructing PageObjects."
+    """Page Factory class for constructing PageObjects.
+    
+    """
 
     @staticmethod
     def create_page(page_object_class_or_interface, \
                     webdriver=None, **kwargs):
         """
         Instantiate a page object from a given Interface or Abstract class.
+
+        Args:
+            page_object_class_or_interface (Class): PageObject class, AbstractBaseClass, or 
+            Interface to attempt to consturct.
         
-        Instantiating a Page from PageObject from class usage:
+        Kwargs:
+            webdriver (WebDriver): Selenium Webdriver to use to instantiate the page.
+        
+        Returns:
+            PageObject
+        
+        Raises:
+            NoMatchingPageError
+        
+        Instantiating a Page from PageObject from class usage::
+        
             my_page_instance = PageFactory.create_page(MyPageClass)
+            
         
-        Instantiating a Page from an Interface or base class
-            import pages.mysite.* 
+        Instantiating a Page from an Interface or base class::
+        
+            import pages.mysite.*  # Make sure you import classes first, or else PageFactory will not know about it.
             my_page_instance = PageFactory.create_page(MyPageInterfaceClass)
+            
         
-        Instantiating a Page from a list of classes.
+        Instantiating a Page from a list of classes.::
+        
             my_page_instance = PageFactory.create_page([PossiblePage1, PossiblePage2])
+            
         
         Note: It'll only be able to detect pages that are imported.  To it's best to 
         do an import of all pages implementing a base class or the interface inside the 
         __init__.py of the package directory.  
         
-        @param  page_object_class_or_interface: Class, AbstractBaseClass, or Interface to attempt to consturct.
-        @param webdriver: Selenium Webdriver to use to instantiate the page.
-        @type webdriver: WebDriver
         """
         
         if not webdriver:
@@ -277,18 +319,24 @@ class PageObjectUtils():
 
     @staticmethod
     def check_css_selectors(webdriver, *selectors):
-        """
-        Returns true if all CSS selectors passed in is found.  This can be used 
+        """Returns true if all CSS selectors passed in is found.  This can be used 
         to quickly validate a page.
         
-        Usage Example:
-        if not PageObjectUtils.check_css_selectors("form#loginForm", "submit.login"):
-            raise InvalidPageError("This is not the login page.")
+        Args:
+            webdriver (Webdriver) : Selenium Webdriver instance
+            selectors (str) : N number of CSS selectors strings to match against the page.
+        
+        Returns:
+            True, False - if the page matches all selectors.
+        
+        Usage Example::
+        
+            # Checks for a Form with id='loginForm' and a button with class 'login'
+            if not PageObjectUtils.check_css_selectors("form#loginForm", "button.login"):
+                raise InvalidPageError("This is not the login page.")
 
-        @param webdriver: WebDriver.
-        @type webdriver: WebDriver 
-        @param *selectors: CSS selector for element to look for.
-        @type *selectors: str
+        You can use this within a PageObject's `_validate_page(webdriver)` method for 
+        validating pages.
         """
         for selector in selectors:
             try:
@@ -301,8 +349,7 @@ class PageObjectUtils():
 
 
 class PageUtils():
-    '''
-    Offers utility methods that operate on a page level.
+    '''Offers utility methods that operate on a page level.
     '''
     
     @staticmethod
@@ -314,13 +361,31 @@ class PageUtils():
         """
         Waits until the page is loaded.
         
-        Usage Example: 
+        Args:
+            page_obj_class (Class) : PageObject class
+        
+        Kwargs:
+            webdriver (Webdriver) : Selenium Webdriver.  Default uses WTF_WEBDRIVER_MANAGER's instance.
+            timeout (number) : Number of seconds to wait to allow the page to load.
+            sleep (number) : Number of seconds to wait between polling.
+            bad_page_classes (list) : List of PageObject classes to fail if matched.  For example, ServerError page.
+        
+        Returns:
+            PageObject
+        
+        Raises:
+            PageUtilOperationTimeoutError : Timeout occurred before the desired PageObject was matched.
+            BadPageEncounteredError : One or more of the PageObject in the specified 'bad_page_classes' list 
+            was matched.
+            
+        
+        Usage Example:: 
             webdriver.get("http://www.mysite.com/login")
             # Wait up to 60 seconds for the page to load.
-            login_page = wait_until_page_loaded(LoginPage, timeout=60)
+            login_page = wait_until_page_loaded(LoginPage, timeout=60, [ServerErrorPage])
         
-        @return: Returns PageObject of type passed in.
-        @rtype: PageObject
+        This will wait for the login_page to load, then return a LoginPage() PageObject.
+
         """
         if not webdriver:
             webdriver = WTF_WEBDRIVER_MANAGER.get_driver()
@@ -360,10 +425,17 @@ class PageUtils():
 
 
     @staticmethod
-    def wait_until_page_ready(pageobject, timeout=WTF_TIMEOUT_MANAGER.NORMAL):
-        "Waits until document.readyState == Complete"
+    def wait_until_page_ready(page_object, timeout=WTF_TIMEOUT_MANAGER.NORMAL):
+        """Waits until document.readyState == Complete (e.g. ready to execute javascript commands)
+        
+        Args:
+            page_object (PageObject) : PageObject class
+        
+        Kwargs:
+            timeout (number) : timeout period
+        """
         try:
-            do_until(lambda: pageobject.webdriver.execute_script("return document.readyState").lower() \
+            do_until(lambda: page_object.webdriver.execute_script("return document.readyState").lower() \
                      =='complete', timeout)
         except wait_utils.OperationTimeoutError:
             raise PageUtilOperationTimeoutError("Timeout occurred while waiting for page to be ready.")
@@ -393,6 +465,7 @@ class PageUtilOperationTimeoutError(Exception):
 class BadPageEncounteredError(Exception):
     "Raised when a bad page is encountered."
     pass
+
 
 class PageLoadTimeoutError(PageUtilOperationTimeoutError):
     "Timeout while waiting for page to load."