+++ wtframework/wtf/web/web_utils.py
@@ -14,14 +14,16 @@
 #    You should have received a copy of the GNU General Public License
 #    along with WTFramework.  If not, see <http://www.gnu.org/licenses/>.
 ##########################################################################
-from urllib2 import urlopen
-import urllib2
-import re
-from wtframework.wtf.web.page import PageFactory
-from wtframework.wtf.config import WTF_TIMEOUT_MANAGER
-import time
-from threading import Thread
 from datetime import datetime, timedelta
+from threading import Thread
+from urllib2 import urlopen
+from wtframework.wtf.config import WTF_TIMEOUT_MANAGER
+from wtframework.wtf.utils.test_utils import do_and_ignore
+from wtframework.wtf.web.page import PageFactory
+from wtframework.wtf.web.webdriver import WTF_WEBDRIVER_MANAGER
+import re
+import time
+import urllib2
 
 class WebUtils(object):
 
@@ -100,7 +102,7 @@ class WebUtils(object):
         
         webdriver.switch_to_window(original_window)
         raise WindowNotFoundError("Window {0} not found.")
-    
+
 
 class BrowserStandBy(object):
     """
@@ -108,10 +110,13 @@ class BrowserStandBy(object):
     a selenium session from timing out.
     """
     
-    def __init__(self, webdriver, max_time=WTF_TIMEOUT_MANAGER.EPIC, sleep=5):
+    def __init__(self, webdriver=None, max_time=WTF_TIMEOUT_MANAGER.EPIC, sleep=5):
         """
         @param webdriver:Webdriver instance to keep alive. 
         """
+        if webdriver is None:
+            webdriver = WTF_WEBDRIVER_MANAGER.get_driver()
+
         self.webdriver = webdriver
         self._sleep_time = sleep
         self._max_time = max_time
@@ -127,19 +132,26 @@ class BrowserStandBy(object):
         self._thread = Thread(target=lambda: self.__stand_by_loop())
         self._keep_running = True
         self._thread.start()
-        pass
-    
+        return self
+
+
     def stop(self):
         "Stop BrowserStandBy from sending additional calls to webdriver."
         self._keep_running = False
-        pass
-    
-    
+        return self
+
+
     def __stand_by_loop(self):
         print self._keep_running
         while datetime.now() < self._end_time and self._keep_running:
             self.webdriver.current_url #Just performing current_url to keep this alive.
             time.sleep(self._sleep_time)
+
+    def __del__(self):
+        do_and_ignore(lambda: self.stop())
+        self._thread = None
+
+
 
 
 class WindowNotFoundError(RuntimeError):